refactor: move figure functions to the single modules
